:rocket: Split up getInitalData to be more reusable
